third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/12.10.1-expected.txt

PASS 'use strict'; with (1) {} threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f() { \"use strict\"; with (1) {} }") threw exception of type SyntaxError.
PASS Function("function f() { \"use strict\"; with (1) {} }") threw exception of type SyntaxError.
PASS true === true
PASS !!Function("function f() { \"use strict\"; }; with (1) {}") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
